refactor(tag): export TAG_TYPE from components
