+++ src/components/tag/tag-story-angular.ts
@@ -11,8 +11,7 @@ import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { storiesOf } from '@storybook/angular';
 import { action } from '@storybook/addon-actions';
 import { withKnobs, boolean, select } from '@storybook/addon-knobs/angular';
-import TAG_TYPE from './types';
-import './tag';
+import { TAG_TYPE } from './tag';
 import './filter-tag';
 
 const createDefaultProps = () => ({
