+++ src/components/tag/tag-story-vue.ts
@@ -11,8 +11,7 @@ import { storiesOf } from '@storybook/vue';
 import { action } from '@storybook/addon-actions';
 import { withKnobs, boolean, select } from '@storybook/addon-knobs';
 import createVueBindingsFromProps from '../../../.storybook/vue/create-vue-bindings-from-props';
-import TAG_TYPE from './types';
-import './tag';
+import { TAG_TYPE } from './tag';
 import './filter-tag';
 
 const createDefaultProps = () => ({
